+++ app/src/main/java/edu/uw/tcss450/chatapp/ui/home/HomeViewModel.java
@@ -1,4 +1,4 @@
-package edu.uw.tcss450.chatapp.ui;
+package edu.uw.tcss450.chatapp.ui.home;
 
 import android.app.Application;
