+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.uw.tcss450.chatapp.ui.HomeFragment">
+    tools:context="edu.uw.tcss450.chatapp.ui.home.HomeFragment">
 
     <TextView
         android:id="@+id/home_weather"
